More refactoring
refers to #7
